Implemented comparison

+++ ExamOOP/ExamOOP/ExamOOP/Date.h
@@ -5,7 +5,7 @@ class Date : public Word
 public:
 	Date();
 	Date(const String & value);
-	int compare(const Word & other) const override;
+	float compare(const Word & other) const override;
 	Word * get_copy() const override;
 };
